debug

+++ common.c
@@ -30,7 +30,7 @@ union ibv_gid get_gid(struct ibv_context *context)
 		fprintf(stderr, "could not get gid for port %d, index %d\n", IB_PHYS_PORT, 0);
 	}
 	fprintf(stderr, "GID: Interface id = %lld subnet prefix = %lld\n", 
-		(long long) ret_gid.global.interface_id, 
+		ret_gid.global.interface_id, 
 		(long long) ret_gid.global.subnet_prefix);
 	
 	return ret_gid;
